Unnecessary comment clean up Notes below regard to ORM vs Nested
NOTE1: ORM satisfies a linked system where Controller objects are linked bidirectional

NOTE2: If one desires to retrieve such as Account data, response type should be <Account>

NOTE3: Account model has a structure where Account has Transactions array and each Transaction item has Account. This is an expected behavior since it is not "nested" but "object-relational"

NOTE4: This also helps to access the Account that Transaction is connected without parent-child logic. They are pre-structured as connected

NOTE5: If one only desires to retrieve such as Account information, it is advised to return AccountDTO where it is an object with desired attributes. It can be customized and diversified